refactor(workerpool): deprecate WorkerPool v1, rename v2 to standard
- Delete WorkerPool v1 implementation (workerPool.ts)
- Rename WorkerPoolV2 to WorkerPool (file and class name)
- Rename workerPoolV2.test.ts to workerPool.test.ts
- Remove WORKER_POOL_V2_ENABLED config option
- Remove WORKER_POOL_SIZE config option (legacy)
- Update all imports and references
- Update tests to remove v2 conditional checks
- Update README environment variables table
- Update workerPoolInterface comments
- Simplify index.ts worker pool initialization

WorkerPool v2 is now the sole implementation with dynamic scaling.
All 149 tests pass.

implement #39
