+++ README.md
@@ -41,14 +41,12 @@ Configure the bot's behavior with the following environment variables:
 | `LOG_LEVEL`                          |    No    | Sets the minimum log level. <br>**Available levels:** `none`, `debug`, `info`, `warn`, `error`, `fatal`.  | `debug` (development)<br>`info` (production) |
 | `LOG_TEMPLATE`                       |    No    | Customizes the log output format.                                                                         | `[{level}: {module}]: {message}`             |
 | `NODE_ENV`                           |    No    | Set to `production` to default the log level to `info`.                                                   | —                                            |
-| `WORKER_POOL_SIZE`                   |    No    | Sets the number of worker threads for regex processing (legacy pool).                                     | 4                                            |
 | `WORKER_TIMEOUT_MS`                  |    No    | Maximum time a regex operation can run before being terminated (milliseconds).                            | 60000                                        |
-| `WORKER_POOL_V2_ENABLED`             |    No    | Enable WorkerPoolV2 with dynamic scaling and health monitoring.                                           | `false`                                      |
-| `WORKER_POOL_MIN_WORKERS`            |    No    | Minimum number of workers to maintain (WorkerPoolV2).                                                     | 0                                            |
-| `WORKER_POOL_MAX_WORKERS`            |    No    | Maximum number of workers allowed (WorkerPoolV2).                                                         | 8                                            |
-| `WORKER_POOL_INITIAL_WORKERS`        |    No    | Number of workers to spawn at startup (WorkerPoolV2).                                                     | 1                                            |
-| `WORKER_POOL_IDLE_TIMEOUT_MS`        |    No    | Time before idle workers are terminated (WorkerPoolV2).                                                   | 300000                                       |
-| `WORKER_POOL_IDLE_CHECK_INTERVAL_MS` |    No    | How often to check for idle workers (WorkerPoolV2).                                                       | 60000                                        |
+| `WORKER_POOL_MIN_WORKERS`            |    No    | Minimum number of workers to maintain.                                                                    | 0                                            |
+| `WORKER_POOL_MAX_WORKERS`            |    No    | Maximum number of workers allowed.                                                                        | 8                                            |
+| `WORKER_POOL_INITIAL_WORKERS`        |    No    | Number of workers to spawn at startup.                                                                    | 1                                            |
+| `WORKER_POOL_IDLE_TIMEOUT_MS`        |    No    | Time before idle workers are terminated.                                                                  | 300000                                       |
+| `WORKER_POOL_IDLE_CHECK_INTERVAL_MS` |    No    | How often to check for idle workers.                                                                      | 60000                                        |
 | `GRACEFUL_DRAIN`                     |    No    | Enable graceful drain on shutdown. Processes pending tasks instead of rejecting them.                     | `false`                                      |
 | `GRACEFUL_DRAIN_TIMEOUT_MS`          |    No    | Maximum time to spend draining queue during shutdown (milliseconds). Max 9500ms for Docker compatibility. | 8000                                         |
 | `MAX_CHAIN_LENGTH`                   |    No    | Maximum number of sed commands that can be chained together.                                              | 5                                            |
