+++ src/index.ts
@@ -14,7 +14,6 @@ import { SED_PATTERN } from "./utils";
 import { DatabaseService } from "./database";
 import { WorkerPool } from "./workerPool";
 import { parseSedCommands, SedHandler } from "./sed";
-import { WorkerPoolV2 } from "./workerPoolV2";
 import type { IWorkerPool } from "./workerPoolInterface";
 import { join } from "node:path";
 import { fileURLToPath } from "node:url";
@@ -24,9 +23,7 @@ import { RateLimitError, TelegramAPIError } from "./errors";
 const {
 	TOKEN,
 	BASE_URL,
-	WORKER_POOL_SIZE,
 	WORKER_TIMEOUT_MS,
-	WORKER_POOL_V2_ENABLED,
 	WORKER_POOL_MIN_WORKERS,
 	WORKER_POOL_MAX_WORKERS,
 	WORKER_POOL_INITIAL_WORKERS,
@@ -167,22 +164,18 @@ const dbService = new DatabaseService(db);
 const __filename = fileURLToPath(import.meta.url);
 const workerScriptPath = join(__filename, "..", "hellspawn.ts");
 
-// Use WorkerPoolV2 if enabled, otherwise use legacy WorkerPool
-const workerPool: IWorkerPool = WORKER_POOL_V2_ENABLED
-	? new WorkerPoolV2({
-			maxWorkers: WORKER_POOL_MAX_WORKERS,
-			minWorkers: WORKER_POOL_MIN_WORKERS,
-			initialWorkers: WORKER_POOL_INITIAL_WORKERS,
-			taskTimeoutMs: WORKER_TIMEOUT_MS,
-			idleTimeoutMs: WORKER_POOL_IDLE_TIMEOUT_MS,
-			idleCheckIntervalMs: WORKER_POOL_IDLE_CHECK_INTERVAL_MS,
-			workerScript: workerScriptPath,
-		})
-	: new WorkerPool(WORKER_POOL_SIZE, workerScriptPath);
+// Initialize WorkerPool with dynamic scaling
+const workerPool: IWorkerPool = new WorkerPool({
+	maxWorkers: WORKER_POOL_MAX_WORKERS,
+	minWorkers: WORKER_POOL_MIN_WORKERS,
+	initialWorkers: WORKER_POOL_INITIAL_WORKERS,
+	taskTimeoutMs: WORKER_TIMEOUT_MS,
+	idleTimeoutMs: WORKER_POOL_IDLE_TIMEOUT_MS,
+	idleCheckIntervalMs: WORKER_POOL_IDLE_CHECK_INTERVAL_MS,
+	workerScript: workerScriptPath,
+});
 
-if (WORKER_POOL_V2_ENABLED) {
-	logger.info("Using WorkerPoolV2 with dynamic scaling");
-}
+logger.info("Using WorkerPool with dynamic scaling");
 
 // --- Sed Handler Setup ---
 async function sendOrEditReply(
